% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formula.R
\name{formula}
\alias{formula}
\title{Get compound formula}
\usage{
formula(chem, representation = "formula")
}
\arguments{
\item{chem, }{chemical name of the compound}

\item{representation, }{representation methods, formula is default}
}
\description{
get compound formula from https://cactus.nci.nih.gov/chemical/structure
}
\examples{
formula('malic acid')
}
