% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formula.R
\name{get_formula}
\alias{get_formula}
\title{Get compound formula and structure}
\usage{
get_formula(chem, representation = "formula", info = FALSE)
}
\arguments{
\item{chem, }{chemical name of the compound}

\item{representation, }{representation methods, formula is default}

\item{info, }{extra molecular information that users can query}
}
\description{
get compound formula and structure from https://cactus.nci.nih.gov/chemical/structure
}
\examples{
get_formula('malic acid')
get_formula(c('malic acid', 'citric acid', 'tartaric acid'))
}
\author{
Yonghui Dong
}
