% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{get_smiles}
\alias{get_smiles}
\title{Get compound smiles}
\usage{
get_smiles(chem)
}
\arguments{
\item{chem, }{chemical name of the compound}
}
\description{
get compound formula and structure from https://cactus.nci.nih.gov/chemical/structure
}
\examples{
get_smiles('malic acid')
get_smiles(c('malic acid', 'citric acid', 'tartaric acid'))
}
\author{
Yonghui Dong
}
