% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Iso_pattern.R
\name{iso_pattern}
\alias{iso_pattern}
\title{Isotope pattern distribution}
\usage{
iso_pattern(F, threshold = 0.01, z, rel_to = 0,
  return_iso_calc_amount = FALSE)
}
\arguments{
\item{F}{chemical formula, must be written in capital letters.}

\item{threshold}{only display the m/z values with relative abundance over the threshold,
default value = 0.01.}

\item{z}{charge.}

\item{rel_to}{0: relative to highest m/z; 1: relative to monoisotopic peak.}

\item{return_iso_calc_amount}{???}
}
\description{
Calculate isotope pattern distribution of given mass formula(e).
}
\note{
This script is adapted from R package 'enviPat', all the credits go to the original author
}
\examples{
 iso_pattern('C7H7O4',z = 1)
 iso_pattern(c('C7H7O4', 'C4H7O6'), z = c(1, -1))
 library(plotly)
 out = iso_pattern('C700H7O4K30',z = 1)
 out = as.data.frame(out$C700H7O4K30)
 plot_ly(out) \%>\% add_trace(x = out$`m/z`, y = out$abundance, type = 'bar', width = .001)
}
