% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypca.R
\name{mypca}
\alias{mypca}
\title{fast PCA}
\usage{
mypca(xset, centering = T, scaling = "none", ms2.rm = FALSE, x = 1,
  y = 2, size = 1.5, exclude = NULL, scale_group = NULL,
  scale_factor = 1, interactive = T, ...)
}
\arguments{
\item{xset}{xcms object}

\item{centering}{centering, default = TRUE}

\item{scaling}{scaling method, default is scaling = "none". You can choose "auto" or "pareto"}

\item{ms2.rm}{remove MSMS data when it is included.}

\item{x}{PCA X axis, default is PC1}

\item{y}{PCA Y axis, defult is PC2}

\item{size}{dot size}

\item{exclude}{exclude some classes of samples}

\item{scale_group}{select groups needs to be scaled.}

\item{scale_factor}{the scale factor, default = 1.}

\item{interactive}{should interactive figure be plotted? default = TRUE. If you want to save the result
in high resolution, use non interative plot.}

\item{...}{other parameters}
}
\description{
perform PCA from xcms object
}
\examples{
\dontrun{
getpca <- mypca(xset)
}
}
\author{
Yonghui Dong
}
