% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysummary.R
\name{mysummary}
\alias{mysummary}
\title{Prefilter}
\usage{
mysummary(xset, ms2.rm, subgroup, scale_group, scale_factor)
}
\arguments{
\item{xset}{xcms object.}

\item{ms2.rm}{should remove MSMS data when it is included, default = TRUE}

\item{subgroup}{subset only specific groups for statistical test, default = NULL, which means statistics will be performed to all group pairs.}

\item{scale_group}{select groups needs to be scaled up or down, accoding to sample dilution or concentration.}

\item{scale_factor}{the scale factor, default = 1.}
}
\value{
a filtered peaklist
}
\description{
prefiltering isotopically labeled analytes according to the experiment design.
}
\examples{
\dontrun{
my_summary <- mysummary(xset)
}
}
