% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_correct.R
\name{RT_correction}
\alias{RT_correction}
\title{Retention time calibration}
\usage{
RT_correction(xset, ppm = 10, rt = 50, anchors = NULL,
  use.DB = "LIPIDOMICS", mode = "+")
}
\arguments{
\item{xset}{xcms object}

\item{ppm}{mass tolerance, default value = 10}

\item{rt}{retention time search window (in second), default value = 50 s.}

\item{anchors}{lipidomics or metabolomics reference peaks}

\item{use.DB}{which database is used for peak identification, choose between "Lipidomics" and "Metabolomics".}

\item{mode}{ionization mode, positive "+" or negative "-".}
}
\description{
calibrate RT according to reference peaks
}
\examples{
\dontrun{
RT.DB <- RT_correction(xset, use.DB = "Lipidomics", mode = "+")
}
}
\author{
Yonghui Dong
}
