% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tanimoto.R
\name{tanimoto}
\alias{tanimoto}
\title{tanimoto similarity}
\usage{
tanimoto(x)
}
\arguments{
\item{x}{similary}
}
\description{
calculate tanimoto similarities of different compounds
}
\author{
Yonghui Dong
}
