% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what.R
\name{what}
\alias{what}
\title{search for m/z in home database}
\usage{
what(mz, mode = c("+", "-"), ppm = 10, relatives = FALSE)
}
\arguments{
\item{mz}{m/z value}

\item{mode}{ionization mode, either positive '+' or negative '-'}

\item{ppm}{mass tolerance, default value = 10}

\item{relatives, }{shuould the corresponding adducts, fragments and parent ions shown}
}
\description{
tentative metabolite identification based on m/z value search in a home database
}
\examples{
 what(133.014, ppm = 10, mode = '-')
 what(133.014, ppm = 10, mode = '-', relative = TRUE)
}
\author{
Yonghui Dong
}
