% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysummary.R
\name{mysummary}
\alias{mysummary}
\title{Prefilter}
\usage{
mysummary(xset, ms2.rm = ms2.rm, subgroup = subgroup)
}
\arguments{
\item{xset}{xcms object.}

\item{ms2.rm}{remove MSMS data when it is included, default = TRUE}

\item{subgroup}{subset the xcms groups. The name should be the same as in phboData$class. default = NULL, which means no subset will be performed.}
}
\value{
a filtered peaklist
}
\description{
prefiltering isotopically labeled analytes according to the experiment design.
}
\examples{
\dontrun{
my_summary <- mysummary(xset)
}
}
