% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_correct.R
\name{RT_correction}
\alias{RT_correction}
\title{Retention time calibration}
\usage{
RT_correction(xset, ppm = 10, rt = 50, anchors = "LIPIDOMICS")
}
\arguments{
\item{xset}{xcms object}

\item{ppm}{mass tolerance, default value = 10}

\item{rt}{retention time search window, default value = 50}

\item{anchors}{lipidomics or metabolomics reference peaks}
}
\description{
calibrate RT according to reference peaks
}
\examples{
 a = 1+1
}
\author{
Yonghui Dong
}
