% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what.R
\name{what}
\alias{what}
\title{search for m/z in home database}
\usage{
what(mz, mode = c("+", "-"), ppm = 10, parent = TRUE)
}
\arguments{
\item{mz}{m/z value}

\item{mode}{ionization mode, either positive '+' or negative '-'}

\item{ppm}{mass tolerance, default value = 10}

\item{parent, }{if parent, adducts and fragments should shown}
}
\description{
tentative metabolite identification based on m/z value search in a home database
}
\examples{
 what(133.014, ppm = 10, mode = '-')
 what(133.014, ppm = 10, mode = '-', parent = FALSE)
}
\author{
Yonghui Dong
}
